Rearrange files to remove circular dependency
